Add second ajax request for product review. Rewrite ajax func into class based
Now users can write reviews for products. For each product each user can write only 1 review it was achived with unique constrain.
Add ajax request for it in order to not refresh the page when user submit form. Add validations for the form.
Rewrite ajax func into class based one and create a decorator in order to check if request is ajax