Updade calculate_bill(). Add check_availability(). Remake discounts
Now calculate_bill() method will take into account all product discount and choose and best one matching the quantity. Also now it takes into account date of the discount only ongoing will be taken into account.
Add check_availability() method will check if cart product is still available if not it will move it to the save for later table. If it is already exists there it won't make anything.
cart_item_instance will be displayed in the template in order to show user if product was moved to the save for lated table and it's not available
Add get_discounts() method in the custom ProductDiscount manager. This method is going to take all discounts for either particular product or just all discounts which will be associated with particular product in the prefetch.
Fix reviews displaying success and fail message.
In the Cart template bill will now either be displayed with discoun or just a basic price. In the future still need to implement order price and logic.