Add forms with ajax request for cart template, also for product and product product_overview
Now user can choose quantity of product in the cart 20 tops. Also price will be updating automatically. There is a bug when there are 1 product with discount and other with main price can be shown wrong. Will be fixed later.
Added ajax functions for removing product from cart, moving it to the shop-for-later table, moving product to the cart from shop-for-later and removing product from shop-for-later.
Added get_best_discount_price in order to dynimically change product price depending on the quantity.
Add ajax request for adding product to the cart in the ajax product snippet. Used it in the products and product_overview templates.
Replace prefetching with annotation in the get_cart_products manager method in order to get the best price depending on the quantity.