Move review creation view to the review app and fix ajax for handling it
Move review for creation review to the review app and add let keywork to the form in the ajax request which caused an error and ajax request didn't work.
Add cart_items variable to all views and in the nav-bar user now can see how many product does he have in the cart.