Add signals for user, product, category, brand.Fix select ordering in js
Now when user sign-in it will automatically be created a profile, wishlist and a cart.
Also for product, category and brand it will automatically be created a slug if it wasn't created yet
Fix select ordering function in js.
Add a dropdown for user if he's autenticated in the navbar for like a shortcut to the wishlist and ability logout.